* commit 'c18822de1291c5a0bdc2eb7434f435b4588038af':
  AArch64: Use long to store pointer

+++ pico/lib/picoctrl.c
@@ -508,10 +508,10 @@ typedef struct picoctrl_engine {
 #define MAGIC_MASK 0x5069436F  /* PiCo */
 
 #define SET_MAGIC_NUMBER(eng) \
-    (eng)->magic = ((picoos_uint32) (eng)) ^ MAGIC_MASK
+    (eng)->magic = ((picoos_uint32) (uintptr_t) (eng)) ^ MAGIC_MASK
 
 #define CHECK_MAGIC_NUMBER(eng) \
-    ((eng)->magic == (((picoos_uint32) (eng)) ^ MAGIC_MASK))
+    ((eng)->magic == (((picoos_uint32) (uintptr_t) (eng)) ^ MAGIC_MASK))
 
 /**
  * performs an engine reset
